+++ clients/admin-ui/src/features/common/table/cells.tsx
@@ -6,6 +6,7 @@ import {
   Tag,
   Text,
   useDisclosure,
+  useToast,
   WarningIcon,
 } from "@fidesui/react";
 import ConfirmationModal from "common/ConfirmationModal";
@@ -13,6 +14,10 @@ import React, { ChangeEvent } from "react";
 import { CellProps } from "react-table";
 
 import ClipboardButton from "~/features/common/ClipboardButton";
+import { getErrorMessage, isErrorResult } from "~/features/common/helpers";
+import { RTKResult } from "~/types/errors";
+
+import { errorToastParams } from "../toast";
 
 export const TitleCell = <T extends object>({
   value,
@@ -25,6 +30,14 @@ export const TitleCell = <T extends object>({
 export const WrappedCell = <T extends object>({
   value,
 }: CellProps<T, string>) => <Text whiteSpace="normal">{value}</Text>;
+
+export const PaddedCell = <T extends object>({
+  value,
+}: CellProps<T, string>) => (
+  <Text whiteSpace="normal" p={2}>
+    {value}
+  </Text>
+);
 
 export const ClipboardCell = <T extends object>({
   value,
@@ -91,7 +104,7 @@ export const MultiTagCell = <T extends object>({
 };
 
 type EnableCellProps<T extends object> = CellProps<T, boolean> & {
-  onToggle: (data: boolean) => Promise<any>;
+  onToggle: (data: boolean) => Promise<RTKResult>;
   title: string;
   message: string;
   /**
@@ -110,8 +123,12 @@ export const EnableCell = <T extends object>({
   isDisabled,
 }: EnableCellProps<T>) => {
   const modal = useDisclosure();
+  const toast = useToast();
   const handlePatch = async ({ enable }: { enable: boolean }) => {
-    await onToggle(enable);
+    const result = await onToggle(enable);
+    if (isErrorResult(result)) {
+      toast(errorToastParams(getErrorMessage(result.error)));
+    }
   };
 
   const handleToggle = async (event: ChangeEvent<HTMLInputElement>) => {